refactor: rename all source files to lower case
