+++ src/kgm/kgm_crypto_type3.cpp
@@ -1,5 +1,5 @@
 #include "kgm_crypto.h"
-#include "utils/LoopIterator.h"
+#include "utils/loop_iterator.h"
 #include "utils/md5.h"
 #include <algorithm>
 #include <cstdint>
@@ -47,8 +47,8 @@ class KGMCryptoType3 final : public IKGMCrypto
 
     template <bool IS_ENCRYPT> void EncryptDecrypt(uint64_t offset, uint8_t *buffer, size_t len)
     {
-        LoopIterator slot_key{slot_key_.data(), slot_key_.size(), offset};
-        LoopIterator file_key{file_key_.data(), file_key_.size(), offset};
+        loop_iterator slot_key{slot_key_.data(), slot_key_.size(), offset};
+        loop_iterator file_key{file_key_.data(), file_key_.size(), offset};
 
         auto *end = buffer + len;
         for (auto *it = buffer; it < end; it++)
