+++ src/ncm/ncm.cpp
@@ -3,10 +3,10 @@
 #include "parakeet-crypto/IStream.h"
 #include "parakeet-crypto/ITransformer.h"
 #include "parakeet-crypto/transformer/ncm.h"
-#include "utils/EndianHelper.h"
-#include "utils/PagedReader.h"
-#include "utils/SizedBlockReader.h"
-#include "utils/XorHelper.h"
+#include "sized_block_reader.h"
+#include "utils/endian_helper.h"
+#include "utils/paged_reader.h"
+#include "utils/xor_helper.h"
 
 #include <algorithm>
 #include <array>
