+++ src/ncm/sized_block_reader.h
