+++ src/qmc1/qmc1_decryption_transformer.h
@@ -1,7 +1,7 @@
 
 #include "parakeet-crypto/ITransformer.h"
-#include "utils/PagedReader.h"
-#include "utils/XorHelper.h"
+#include "utils/paged_reader.h"
+#include "utils/xor_helper.h"
 
 #include <algorithm>
 #include <array>
