+++ src/utils/xor_helper.h
