+++ src/ximalaya/xmly_decryption.cpp
@@ -1,7 +1,7 @@
 #include "parakeet-crypto/ITransformer.h"
 #include "parakeet-crypto/transformer/ximalaya.h"
-#include "utils/PagedReader.h"
-#include "utils/XorHelper.h"
+#include "utils/paged_reader.h"
+#include "utils/xor_helper.h"
 #include <algorithm>
 #include <array>
 #include <cstdint>
